New current velocity is made available for torque calculation
Torque variances that are proportional to velocity (such as friction &
back-EMF) are now calculated based on the new velocity calculated during
a pose update, not the previous velocity. This ensures that the
acceleration value integrated during the next update fully reflects the
current velocity integrated into position during the same next update.