DC perm. mag. motor torque est. now respects `plugging` back EMF boost
When a DC permanent magnet motor is rotating and is commanded to apply
force in the opposite direction, the max available torque is increased
by the current speed of the motor rather than decreased. This is due to
the fact that the back EMF voltage generated by a motor supports
applied voltage in this case rather than cancelling it.

E.g. if a motor is travelling at negative full speed and positive full
duty cycle is applied at full voltage, the motor will generate 2x max
torque, which declines linearly as the motor slows to stall (at which
1x max torque is available) and as it speeds up in the positive direction
(until the motor nears max speed, near which 0x max torque is available).